refactor: split recorder components
